+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to the [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.2.0 -- 2019-06-26
+
+* Add `PrimUnlifted` instances for `MVar`, `IORef`, and `STRef`.
+
+## 0.1.1.0 -- 2019-05-23
+
+* Cannot remember what happened in this release.
+
+## 0.1.0.0 -- 2019-05-17
 
 * First version. Released on an unsuspecting world.